Conformance Test fixes for Windows
* Paths unsupported in load library for python 3.9
* RESCALE min/max calculations not promoted to int64
* Utils product() calculation also not promoted to large enough storage
* Fixed color printing output
* Made conformance generator build ops in sorted order for consistence
* Updated default model file locations and tool help messages
* Fixed TransposePermsOutputShapeMismatch ERROR_IF

+++ verif/generator/tosa_utils.py
@@ -187,7 +187,8 @@ def usableDTypes(*, excludes=None):
 
 
 def product(shape):
-    value = 1
+    # Make sure we use a large enough storage type
+    value = np.int64(1)
     for n in shape:
         value *= n
     return value
